// src/core/api/model/medicalRecordResponseDto.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Dentizy API
 * API Documentation untuk Sistem Manajemen Klinik Gigi
 * OpenAPI spec version: 1.0
 */
import type { AppointmentSubsetDto } from './appointmentSubsetDto';
import type { DoctorSubsetDto } from './doctorSubsetDto';
import type { MedicalRecordResponseDtoAssessment } from './medicalRecordResponseDtoAssessment';
import type { MedicalRecordResponseDtoDeletedAt } from './medicalRecordResponseDtoDeletedAt';
import type { MedicalRecordResponseDtoObjektif } from './medicalRecordResponseDtoObjektif';
import type { MedicalRecordResponseDtoPlan } from './medicalRecordResponseDtoPlan';
import type { MedicalRecordResponseDtoSubjektif } from './medicalRecordResponseDtoSubjektif';
import type { MedicalRecordTreatmentSubsetDto } from './medicalRecordTreatmentSubsetDto';
import type { PatientSubsetDto } from './patientSubsetDto';

export interface MedicalRecordResponseDto {
  /** ID unik rekam medis */
  id: string;
  appointment_id: string;
  doctor_id: string;
  patient_id: string;
  /** @nullable */
  subjektif?: MedicalRecordResponseDtoSubjektif;
  /** @nullable */
  objektif?: MedicalRecordResponseDtoObjektif;
  /** @nullable */
  assessment?: MedicalRecordResponseDtoAssessment;
  /** @nullable */
  plan?: MedicalRecordResponseDtoPlan;
  created_at: string;
  updated_at: string;
  /** @nullable */
  deleted_at?: MedicalRecordResponseDtoDeletedAt;
  umur_rekam: number;
  appointment?: AppointmentSubsetDto;
  doctor?: DoctorSubsetDto;
  patient?: PatientSubsetDto;
  medical_record_treatments?: MedicalRecordTreatmentSubsetDto[];
}